[#29] GroupBox 파일 분리
